Create domain context and generic mapper
Create EFCore DbContext with DbSets for each of the classes in the
domain model. This enables EFCore to create the database schema,
including the necessary join tables.

Create generalized Mapper class that can handle mapping of all classes
in the domain model.
It may be necessary with addtional configuration to handle loading of
related data.