Add optional argument to program
Add the option of specifying the path of the config file as a command
line parameter when starting the program.
If no argument is provided, the program continues to work as before,
using the default path of the configuration file.
Error messages are printed if too many arguments are given, or if the
filepath cannot be opened.